Change the renamer so that it's much more simple in terms of code.
Yes, this doesn't do much, but it makes the branching logic a lot redundant, especially, when we just can transform the int into a string, and measure it's length.
This will properly handle the case of 10^n where n > 3, though I don't think I know of a show with something on the scale of 10k episodes.