Overhaul for the ShokoServer.CLI csproj
TargetFramework: Changed net5 to net5.0 as this is officially used by VS2019
RuntimeIdentifiers: Not used any more in project files instead it can be set at the command line during build. However this is most likely not needed anymore since dotnetcore 1.x
Platforms: Deleted x64. AnyCPU chooses the best option for the hosting system. On x64 Systems this is exactly the same as x64. However this is only a flag in the assembly and can be changed using Microsoft’s CoreFlags tool after the build.
Deterministic: Using deterministic builds is the new default and has many benefits over non deterministic builds.
AutoGenerateBindingRedirects: Just a personal opinion but handling this thing youself ist the more secure way.
ApplicationIcon: db.ico wa linked but never used. However a linkled Item can not be used as ApplicationIcon. I added the relative path to the Icon in the settings so that the icon now is used as application icon.