misc: remove try…catch and force fetch missing xml files
Remove the try…catch in the action, exposing the raw error and stack trace
to both the client and the http logging middleware. And also tweak the
command to queue the missing xml files so it tries to forcefully fetch
them, if needed. It will still try to load from the cache if we're http
banned and the xml file exists.