refactor: consilidate endpoints and update internals
Consilidate the two new v3 endpoints into one by moving the logic from
one of the endpoints into the internals of the other endpoints. Also
move the try…catch into the v1 controller and expose the error to the
v3 clients and let the http logger middleware handle the logging of the
error.

Also fixed up the remaining v3 endpoint by adding comments, and
removed the possible null state of the input by adding a default value
(that's also exposed to and used by swagger in the documentation
generation).