misc: add db migration for setting the timestamps
Add a C# migration path for setting the intended timestamp where
possible, or otherwise resetting the timestamp to make sure we're
able to update the episode entries in the future with newer, correct
data. Any "faulty" episodes will also be reset, and their anime entry
will be queued for update.