fix: fix up episode linking
just a few tweaks here and there, and also fix "unlinking" an episode
by setting the tmdb id to 0 (which was disallowed in the api). setting
the id to 0 instead of deleting the mapping for the shoko/anidb episode
will prevent the auto-matching logic to re-adding new links for the
shoko/anidb episode as long as we don't specify we want to remove all
existing links.