feat: add hidden flag for tmdb episodes
Since I will need it to hide "missing" episodes in a TMDB structure that I never intend to get and don't want to show up in Jellyfin/Shokofin. We already have this for Shoko/AniDB episodes, but in a pure TMDB structured series in Shokofin then we don't use the Shoko/AniDB episodes at all, and most if not all of the episodes that you would want hidden you would probably also not want linked to Shoko/AniDB, so a way to hide them when they're not necessarily linked to Shoko/AniDB was needed.