refactor: reimplement anime↔character/creator mappings for AniDB
Reimplemented the AniDB Character and Creator mappings for AniDB anime. Making sure we store all the mappings, and in the correct order they appear in the XMLs. This is done by dropping the existing tables, followed by creating new ones and restoring the data from the XMLs if possible, or queueing the anime for a forced online lookup if it's not in the local XML cache or is otherwise unusable.

Dropped the `AnimeStaff` and `CrossRef_Anime_Staff` tables used by the API, which were almost a 1:1 mapping for `AniDB_Anime_Staff` combined with `AniDB_Anime_Character`, except sometimes out-of-sync data. All usage of the tables has been directed at the updated `AniDB_Character`, `AniDB_Anime_Staff`, `AniDB_Anime_Character`, or `AniDB_Anime_Character_Creator` tables where appropriate.

Fixes #1183 Closes #1183