fix: set all fields when converting the current TMDB image to `IImageMetadata`
Set all the fields on a cloned object when we convert the TMDB image to an `IImageMetadata` object if the `preferred` value difference from the image's `IsPreferred`, otherwise just return the original image. This will ensure we're sending the correct metadata for the preferred image when it's a TMDB image.