fix: add more fields to AniDB characters
Added parsing of the character type (character, organization, etc.) and last updated timestamp for AniDB characters. The first one will allow us to distinguish between characters and organizations, while the latter will allow us to do conditional updates of the data found in the xml like we do for the episodes and tags, in case you ever attempt to refresh from a partially outdated xml where it has a character that's shared between multiple anime or whatnot. There is also user votes and user ratings to be found in the xmls we can add in the future, if the need ever arise.

Also fixed it so organizations show up in APIv3.

And last; the database migration for SQLite is tested, but for the two other databases it's not. So if somebody can test that they work, and report back on it then that would be great.