fix: only rename/move files if _some_ of the anime data changed
Conditionally only react to some changes of the anidb anime data when scheduling files for the anime to be renamed/moved. The exact number of fields to react to may be fine tuned further in another commit if needed, but this should prevent changes to the number of ratings and such to trigger a full dispatch on all the files related to the anime to be renamed/moved again.