refactor: fix TMDb image relations
Fixed the TMDb image relations to support a proper many-to-many mapping between entities of different types and images. The previous implementation was flawed in that it assumed only a single entity of a given type would be linked to the same image at all times. That is wrong. I'll assume they're using some sort of hashing on their side to automagically de-duplicate exact images between entities, leading to images being re-used between entities at semi-random, but that's just an estimated guess, and I don't actually know how their internals work.