refactor: switch `IQueryable` to `IEnumerable` for enumerable collections in the plugin abstraction
Switched `IQueryable` to `IEnumerable` for enumerable collections exposed through the services in the plugin abstraction because they're not directly tied to any queryable sources that directly use the plugin interfaces. It's better to just advertise them as normal `IEnumerable`s in this case. I'm sorry to all the 4 beta users that need to update their abstraction target once again to accommodate this change.