refactor: add 'is automatic' flag to release info providers
Added a 'is automatic' flag to the new release info request (which also holds the video now) to be submitted to the providers so they can know if it's an automatic or manual search. This may be useful for plugin providers if they want to do conditional logic depending on if the search is manual or automatic. E.g. doing a group based allow/deny list during automatic searches, etc.