refactor: update AniDB service in plugin abstraction & related
Updated the AniDB service in the plugin abstraction to include a proper ban occurred and ban expired event, in addition to storing the last event for each of the ban types for static access. Also reconfigured the APIv3 endpoint to use the new static accessible values instead of going to the connection handlers directly.

Also refactored the internals to properly emit proper ban occurred and ban expires events, instead of implictly relying on the banned time on the connection handler interface combined with the status update events (which can also include other event types). It also created a direct link between the `BanTime` and the `IsBanned` properties on the connection handler.

And last but not least, lowercased some letters in the class/property names related to the AniDB service to fit the overall style of the plugin abstraction.