refactor: add `searchById` query parameter to the AniDB search API in APIv3
Added a new query parameter, `searchById` to the AniDB search endpoints in APIv3. It is enabled by default to match the current behavior, but this allows clients to opt out of the opinionated behavior hiding search results for numbers — i.e. searching for '86' or '91' brings up the AniDB anime entry with the id matching the number instead of searching the series titles for the text.